drafts/extends/requirements.txt
Django
redis
markdown
Pillow
awesome-slugify
